Refactor to eliminate bundle Meteor tool
For a significant number of installations, the Meteor build tool is not
actually required, so to include it inside the container image is
wasteful.  In particular, if you are supplying a pre-bundled Meteor app,
then all that is necessary is Node.

This reorganizes the container such that the Meteor tool is downloaded
at runtime if and only if it is needed to build the bundle.

Signed-off-by: Ulexus <ulexus@gmail.com>